Refactor RemoteApiService with RemoteRequest helper
